✅ Trivial Elm app without `Animation`
👇Difference from previous "No error" commit👇

Simplified `noerror` version to account for (ancient) feedback on github

Basically, I was exhausted from fighting the compilation error at the time I first set up this repo and didn't realize I'd missed a crucial aspect of the error: that it only happens when compiling with the `--debug` flag. I didn't realize that Parcel was compiling with --debug by default, so the problem seemed far more complex to me than it actually is.